[FEAT] stop all clients if signals are received from multiple clients & error handling of write()

+++ src/server.c
@@ -6,11 +6,13 @@
 /*   By: akyoshid <akyoshid@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/28 18:47:02 by akyoshid          #+#    #+#             */
-/*   Updated: 2025/01/30 00:39:52 by akyoshid         ###   ########.fr       */
+/*   Updated: 2025/01/30 02:10:33 by akyoshid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/minitalk.h"
+
+volatile sig_atomic_t	g_err_flag;
 
 void	proc_bit_count_8(unsigned char *c, int *bit_count, pid_t *client_pid)
 {
@@ -23,7 +25,11 @@ void	proc_bit_count_8(unsigned char *c, int *bit_count, pid_t *client_pid)
 	}
 	else
 	{
-		write(1, c, 1);
+		if (write(1, c, 1) == -1)
+		{
+			g_err_flag = ERR_WRITE;
+			return ;
+		}
 		*c = 0;
 		*bit_count = 0;
 	}
@@ -34,17 +40,17 @@ void	handler(int signum, siginfo_t *info, void *context)
 	static unsigned char	c;
 	static int				bit_count;
 	static pid_t			client_pid;
-	// When the client does not terminate correctly, client_pid is not reset,
-	// so the server cannot accept bits from the new client.
 
 	(void)context;
 	if (client_pid == 0)
 		client_pid = info->si_pid;
 	else if (client_pid != info->si_pid)
 	{
+		kill(client_pid, SIGUSR2);
 		kill(info->si_pid, SIGUSR2);
-		// if (client_pid != 0)
-		// 	kill(client_pid, SIGUSR1);
+		c = 0;
+		bit_count = 0;
+		client_pid = 0;
 		return ;
 	}
 	if (signum == SIGUSR2)
@@ -52,6 +58,8 @@ void	handler(int signum, siginfo_t *info, void *context)
 	bit_count++;
 	if (bit_count == 8)
 		proc_bit_count_8(&c, &bit_count, &client_pid);
+	if (g_err_flag == ERR_WRITE)
+		return ;
 	if (client_pid != 0)
 		kill(client_pid, SIGUSR1);
 }
@@ -73,6 +81,8 @@ int	main(void)
 	sigaction(SIGUSR2, &sa, NULL);
 	while (1)
 	{
+		if (g_err_flag == ERR_WRITE)
+			proc_err(ERR_WRITE);
 		if (usleep(100) == 0)
 		{
 			usleep_count++;
